% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncs.R
\name{ssr_get_vectors}
\alias{ssr_get_vectors}
\title{Extract words embeddings from StarSpace model.}
\usage{
ssr_get_vectors(model, words = get_dictionary(model))
}
\arguments{
\item{path}{}
}
\description{
Extract words embeddings from StarSpace model.
}
