% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{StarSpaceR}
\alias{StarSpaceR}
\alias{StarSpaceR-package}
\alias{StarSpaceR-package}
\title{StarSpaceR: 'StarSpace' Wrapper for Text Classification and Word Representation}
\description{
Wrapper for StarSpace library.
}
\author{
\strong{Maintainer}: Zygmunt Zawadzki \email{zygmunt@zstat.pl} [copyright holder]

Other contributors:
\itemize{
  \item Facebook, Inc \email{ledell@fb.com} [copyright holder]
}

}
